moved task list to dashboard, cleanup
