+++ pages/Dashboard/styles.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Task = styled.span`
+  display: block;
+`;
+Task.displayName = "Task";
